refactor: Add param in TransformNumericalVars
When you applyi transf in data, most of the time you use the mean, median or max of the
train.data. In the prediction phase you will need this values. So I add extra arg where I pass the train.data
to recompute all stat used in any transformatio. Temp solution for now

* Remove sub median(x) in TransformVariable (simplification)
* Add arg train.data (It is the ref for the transformations) in TransformNumericalVars to