chore: Add comment to debug BUG in XGBoostExplorer
The BUG is due to vars that are not numerical because xgboost does not know to treat bool and fators

If I convert all vars to numeric, glm predict stop working because it converts all vars as bool or factor and because
predict is not in the same envroment that glm was called, it return an error message that factors level is not the same
in fit and predict

I will not use xgboost for this project. I will use relative importance in the caret package.